+++ src/main/java/gov/ismonnet/client/renderer/swing/WallRenderer.java
@@ -14,9 +14,9 @@ class WallRenderer implements Renderer<SwingRenderContext, WallEntity> {
     public void render(SwingRenderContext ctx, WallEntity toRender) {
         ctx.setColor(Color.yellow);
         ctx.fillRect(
-                (int) toRender.getPosX(),
-                (int) toRender.getPosY(),
-                (int) toRender.getWidth(),
-                (int) toRender.getHeight());
+                toRender.getPosX(),
+                toRender.getPosY(),
+                toRender.getWidth(),
+                toRender.getHeight());
     }
 }
